refactor: move ActivityOptionService from internal/activity to root package
